requirements.txt
Flask==3.0.0
flask-login==0.6.3
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

